!! QR Code Reader
There's a lot of questionable code here. The bulk of the changes help
determine if we should include the extension ane or not (if its android
or ios, we should; otherwise, no) and whether or not the QR code reader
is supported on the current device. Diagnostic data can be found on the
authentication login screen.