UI for reordering assets in a grid
Its not pretty code or display, but it gets the job done. Now I just
need to make the API call to update the asset order in the DB, then
redraw the grid with the new order. Ideally that happens without
unloading and loading the whole exhibit, but I'm not otherwise sure how
to target the parent module. Maybe module.as includes a refresh
function...